Add SqlTableMappingValidator per suggestion in CR comment
