+++ data-jdbc/src/test/groovy/io/micronaut/data/jdbc/mariadb/MariaSchemaSpec.groovy
@@ -4,7 +4,6 @@ import io.micronaut.context.ApplicationContext
 import io.micronaut.data.annotation.Id
 import io.micronaut.data.annotation.MappedEntity
 import io.micronaut.data.connection.jdbc.advice.DelegatingDataSource
-import io.micronaut.data.exceptions.SchemaValidationException
 import io.micronaut.data.jdbc.annotation.JdbcRepository
 import io.micronaut.data.model.query.builder.sql.Dialect
 import io.micronaut.data.repository.CrudRepository
