+++ data-model/src/main/java/io/micronaut/data/model/query/builder/sql/SqlSchemaUtils.java
@@ -28,7 +28,6 @@ import io.micronaut.data.annotation.MappedProperty;
 import io.micronaut.data.annotation.Relation;
 import io.micronaut.data.annotation.sql.SqlMembers;
 import io.micronaut.data.exceptions.MappingException;
-import io.micronaut.data.exceptions.SchemaValidationException;
 import io.micronaut.data.model.Association;
 import io.micronaut.data.model.DataType;
 import io.micronaut.data.model.PersistentEntity;
@@ -40,13 +39,10 @@ import io.micronaut.data.model.schema.sql.SqlColumnMapping;
 import io.micronaut.data.model.schema.sql.SqlDbType;
 import io.micronaut.data.model.schema.sql.SqlSequenceMapping;
 import io.micronaut.data.model.schema.sql.SqlTableMapping;
-import io.micronaut.data.model.schema.sql.metadata.SqlColumnMetadata;
-import io.micronaut.data.model.schema.sql.metadata.SqlTableMetadata;
 
 import java.lang.annotation.Annotation;
 import java.sql.Blob;
 import java.sql.Clob;
-import java.sql.Types;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
@@ -213,207 +209,6 @@ public final class SqlSchemaUtils {
         SqlTableMapping table = new SqlTableMapping(schema, tableName, SqlTableMapping.TableType.MAIN, primaryKeyColumns, columns, sequences);
         tables.add(table);
         return tables;
-    }
-
-    /**
-     * Validates a table definition based on {@link PersistentEntity} mapping against its actual corresponding metadata from the database.
-     *
-     * @param tableMapping    the SQL table mapping from {@link PersistentEntity} to validate
-     * @param tableMetadata   the SQL table metadata from the database to compare against
-     * @param dialect         the SQL dialect of the schema
-     * @throws SchemaValidationException When expected column not found or is not matching expected type
-     */
-    public static void validateTable(@NonNull SqlTableMapping tableMapping, @NonNull SqlTableMetadata tableMetadata, @NonNull Dialect dialect) {
-        List<SqlColumnMapping> primaryKeyColumns = tableMapping.primaryKeyColumns();
-        List<SqlColumnMapping> columns = tableMapping.columns();
-        List<SqlColumnMapping> allColumns = new ArrayList<>(columns.size() + (primaryKeyColumns != null ? primaryKeyColumns.size() : 0));
-        if (primaryKeyColumns != null) {
-            allColumns.addAll(primaryKeyColumns);
-        }
-        allColumns.addAll(columns);
-        for (SqlColumnMapping columnMapping : allColumns) {
-            String name = columnMapping.getName();
-            SqlColumnMetadata columnMetadata = tableMetadata.getColumn(name.toLowerCase());
-            if (columnMetadata == null) {
-                throw new SchemaValidationException("Schema validation failed. Column [" + name + "] not found in the table [" + tableMapping.name() + "]");
-            }
-            validateColumn(columnMapping, columnMetadata, dialect, tableMetadata.getName());
-        }
-    }
-
-    /**
-     * Validates a single column mapping definition against its corresponding metadata from the database.
-     *
-     * @param columnMapping     the SQL column mapping from {@link PersistentEntity} field to validate
-     * @param columnMetadata    the SQL column metadata from the database to compare against
-     * @param dialect           the dialect of the schema where column belongs
-     * @param tableName         the name of the table where column is stored
-     * @throws SchemaValidationException when the expected column does not match the actual column metadata
-     */
-    private static void validateColumn(SqlColumnMapping columnMapping, SqlColumnMetadata columnMetadata,
-                                       Dialect dialect, String tableName) {
-        if (StringUtils.isNotEmpty(columnMapping.getDefinition())) {
-            // Don't compare columns with custom SQL definition
-            // and let user be responsible for mapping of that field
-            return;
-        }
-        if (matchingColumn(columnMapping, columnMetadata, dialect)) {
-            return;
-        }
-        throw new SchemaValidationException(String.format("Schema validation failed. Column [%s] in table [%s] of type [%s] is mapped to [%s].",
-            columnMetadata.name(), tableName, columnMetadata.typeName(), columnMapping.getDbType()));
-    }
-
-    private static boolean matchingColumn(SqlColumnMapping columnMapping, SqlColumnMetadata columnMetadata, Dialect dialect) {
-        if (matchingColumnTypes(columnMapping.getDbType(), columnMetadata.type())) {
-            return true;
-        }
-        String sqlType = columnMapping.getSqlType(dialect);
-        if (sqlType.equalsIgnoreCase(columnMetadata.typeName())) {
-            return true;
-        }
-        if (dialect == Dialect.ORACLE) {
-            return matchOracleColumn(columnMapping, columnMetadata);
-        } else if (dialect == Dialect.MYSQL) {
-            return matchMySqlColumn(columnMapping, columnMetadata);
-        } else if (dialect == Dialect.H2) {
-            return matchH2Column(columnMapping, columnMetadata);
-        } else if (dialect == Dialect.POSTGRES) {
-            return matchPostgresColumn(columnMapping, columnMetadata);
-        } else if (dialect == Dialect.SQL_SERVER) {
-            return matchSqlServerColumn(columnMapping, columnMetadata);
-        }
-        // Add other rules for matching if needed
-        return false;
-    }
-
-    private static boolean matchOracleColumn(SqlColumnMapping columnMapping, SqlColumnMetadata columnMetadata) {
-        if (columnMapping.getDbType() == SqlDbType.UUID) {
-            return uuidMatchesVarchar(columnMetadata);
-        } else if (columnMetadata.type() == Types.NUMERIC) {
-            // Custom sql type name for ORACLE
-            String oracleSqlType = "NUMBER";
-            if (columnMetadata.columnSize() > 0) {
-                oracleSqlType += "(" + columnMetadata.columnSize();
-                if (columnMetadata.decimalDigits() > 0) {
-                    oracleSqlType += "," + columnMetadata.decimalDigits();
-                }
-                oracleSqlType += ")";
-            }
-            return columnMapping.getSqlType(Dialect.ORACLE).equalsIgnoreCase(oracleSqlType);
-        } else if (isOracleBinaryDoubleOrFloat(columnMetadata.typeName())) {
-            return isFloatOrRealOrDouble(columnMapping.getDbType().getType());
-        }
-        return false;
-    }
-
-    private static boolean isOracleBinaryDoubleOrFloat(String typeName) {
-        return "BINARY_DOUBLE".equalsIgnoreCase(typeName) || "BINARY_FLOAT".equalsIgnoreCase(typeName);
-    }
-
-    private static boolean matchMySqlColumn(SqlColumnMapping columnMapping, SqlColumnMetadata columnMetadata) {
-        if (columnMapping.getDbType() == SqlDbType.UUID) {
-            return uuidMatchesVarchar(columnMetadata) ||
-                // For MariaDB
-                (columnMetadata.type() == Types.OTHER && columnMetadata.typeName().equalsIgnoreCase("uuid"));
-        }
-        if (columnMapping.getDbType() == SqlDbType.BOOLEAN) {
-            return columnMetadata.type() == Types.BIT;
-        }
-        if (columnMapping.getDbType() == SqlDbType.JSON) {
-            return columnMetadata.type() == Types.LONGVARCHAR;
-        }
-        return false;
-    }
-
-    private static boolean matchH2Column(SqlColumnMapping columnMapping, SqlColumnMetadata columnMetadata) {
-        if (columnMapping.getDbType() == SqlDbType.BINARY) {
-            return columnMetadata.type() == Types.BLOB;
-        }
-        return false;
-    }
-
-    private static boolean matchPostgresColumn(SqlColumnMapping columnMapping, SqlColumnMetadata columnMetadata) {
-        if (columnMapping.getDbType() == SqlDbType.BOOLEAN) {
-            return columnMetadata.type() == Types.BIT;
-        }
-        return false;
-    }
-
-    private static boolean matchSqlServerColumn(SqlColumnMapping columnMapping, SqlColumnMetadata columnMetadata) {
-        if (columnMapping.getDbType() == SqlDbType.BINARY) {
-            return columnMetadata.type() == Types.VARBINARY;
-        }
-        if (columnMapping.getDbType() == SqlDbType.JSON) {
-            return columnMetadata.type() == Types.NVARCHAR;
-        }
-        return false;
-    }
-
-    private static boolean uuidMatchesVarchar(SqlColumnMetadata columnMetadata) {
-        return columnMetadata.type() == Types.VARCHAR && columnMetadata.columnSize() == 36;
-    }
-
-    private static boolean matchingColumnTypes(SqlDbType dbType, int typeCode) {
-        int mappedTypeCode = dbType.getType();
-        return mappedTypeCode == typeCode
-                || isCompatibleIntegralType(mappedTypeCode, typeCode)
-                || isNumericOrDecimal(mappedTypeCode) && isNumericOrDecimal(typeCode)
-                || isFloatOrRealOrDouble(mappedTypeCode) && isFloatOrRealOrDouble(typeCode)
-                || isVarcharType(mappedTypeCode) && isVarcharType(typeCode)
-                || isVarbinaryType(mappedTypeCode) && isVarbinaryType(typeCode)
-                || isEnumType(mappedTypeCode) && isVarcharType(typeCode);
-        // Add more checks/fallbacks during testing and/or reported issues
-    }
-
-    private static boolean isCompatibleIntegralType(int typeCode1, int typeCode2) {
-        return switch (typeCode1) {
-            case Types.TINYINT ->
-                typeCode2 == Types.TINYINT
-                    || typeCode2 == Types.SMALLINT
-                    || typeCode2 == Types.INTEGER
-                    || typeCode2 == Types.BIGINT;
-            case Types.SMALLINT ->
-                typeCode2 == Types.SMALLINT
-                    || typeCode2 == Types.INTEGER
-                    || typeCode2 == Types.BIGINT;
-            case Types.INTEGER ->
-                typeCode2 == Types.INTEGER
-                    || typeCode2 == Types.BIGINT;
-            default -> false;
-        };
-    }
-
-    private static boolean isNumericOrDecimal(int typeCode) {
-        return switch (typeCode) {
-            case Types.NUMERIC, Types.DECIMAL ->  true;
-            default -> false;
-        };
-    }
-
-    private static boolean isFloatOrRealOrDouble(int typeCode) {
-        return switch (typeCode) {
-            case Types.FLOAT, Types.REAL, Types.DOUBLE -> true;
-            default -> false;
-        };
-    }
-
-    private static boolean isVarcharType(int typeCode) {
-        return switch (typeCode) {
-            case Types.VARCHAR, Types.LONGVARCHAR, Types.NVARCHAR, Types.LONGNVARCHAR -> true;
-            default -> false;
-        };
-    }
-
-    private static boolean isVarbinaryType(int typeCode) {
-        return switch (typeCode) {
-            case Types.VARBINARY, Types.LONGVARBINARY -> true;
-            default -> false;
-        };
-    }
-
-    private static boolean isEnumType(int typeCode) {
-        return typeCode == SqlDbType.ENUM.getType();
     }
 
     /**
