+++ data-model/src/main/java/io/micronaut/data/model/query/builder/sql/validation/BaseSqlTableMappingValidator.java
@@ -1,0 +1,216 @@
+/*
+ * Copyright 2017-2025 original authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.micronaut.data.model.query.builder.sql.validation;
+
+import io.micronaut.core.util.StringUtils;
+import io.micronaut.data.model.PersistentEntity;
+import io.micronaut.data.model.query.builder.sql.Dialect;
+import io.micronaut.data.model.schema.sql.SqlColumnMapping;
+import io.micronaut.data.model.schema.sql.SqlDbType;
+import io.micronaut.data.model.schema.sql.SqlTableMapping;
+import io.micronaut.data.model.schema.sql.metadata.SqlColumnMetadata;
+import io.micronaut.data.model.schema.sql.metadata.SqlTableMetadata;
+
+import java.sql.Types;
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * An abstract base class for validating SQL table mappings against actual table metadata from the database.
+ * <p>
+ * This class provides a basic implementation of the {@link SqlTableMappingValidator} interface,
+ * including validation of table and column definitions extracted from a {@link PersistentEntity}
+ * against the corresponding metadata from the database.
+ * <p>
+ * Subclasses are expected to implement the {@link #getSupportedDialect()} method to specify the SQL dialect
+ * supported by the validator and optionally override {@link #matchingDialectColumnType(SqlColumnMapping, SqlColumnMetadata)}
+ * to provide dialect-specific column type matching logic.
+ *
+ * @since 4.13.0
+ * @author radovanradic
+ */
+abstract class BaseSqlTableMappingValidator implements SqlTableMappingValidator {
+
+    @Override
+    public final void validateTable(SqlTableMapping tableMapping, SqlTableMetadata tableMetadata) {
+        List<SqlColumnMapping> primaryKeyColumns = tableMapping.primaryKeyColumns();
+        List<SqlColumnMapping> columns = tableMapping.columns();
+        List<SqlColumnMapping> allColumns = new ArrayList<>(columns.size() + (primaryKeyColumns != null ? primaryKeyColumns.size() : 0));
+        if (primaryKeyColumns != null) {
+            allColumns.addAll(primaryKeyColumns);
+        }
+        allColumns.addAll(columns);
+        for (SqlColumnMapping columnMapping : allColumns) {
+            String name = columnMapping.getName();
+            SqlColumnMetadata columnMetadata = tableMetadata.getColumn(name.toLowerCase());
+            if (columnMetadata == null) {
+                throw new SchemaValidationException("Schema validation failed. Column [" + name + "] not found in the table [" + tableMapping.name() + "]");
+            }
+            validateColumn(columnMapping, columnMetadata, getSupportedDialect(), tableMetadata.getName());
+        }
+    }
+
+    /**
+     * Checks if the column type defined in the {@link SqlColumnMapping} matches the actual column type
+     * retrieved from the database metadata ({@link SqlColumnMetadata}) for the given SQL dialect.
+     * <p>
+     * The method performs a multi-step comparison:
+     * <ol>
+     *     <li>Compares the database-specific type code from {@link SqlColumnMapping#getDbType()} with the type code from {@link SqlColumnMetadata#type()}.</li>
+     *     <li>If the type codes do not match, it compares the SQL type representation generated by {@link SqlColumnMapping#getSqlType(Dialect)}
+     *         with the type name from {@link SqlColumnMetadata#typeName()}.</li>
+     *     <li>If the above checks fail, it delegates to {@link #matchingDialectColumnType(SqlColumnMapping, SqlColumnMetadata)}
+     *         for dialect-specific matching logic.</li>
+     * </ol>
+     *
+     * @param columnMapping  the SQL column mapping from {@link PersistentEntity} field
+     * @param columnMetadata the SQL column metadata from the database
+     * @param dialect        the SQL dialect to consider during type comparison
+     * @return true if the column type matches, false otherwise
+     */
+    protected final boolean matchingColumnType(SqlColumnMapping columnMapping, SqlColumnMetadata columnMetadata,
+                                               Dialect dialect) {
+        if (matchingColumnTypes(columnMapping.getDbType(), columnMetadata.type())) {
+            return true;
+        }
+        String sqlType = columnMapping.getSqlType(dialect);
+        if (sqlType.equalsIgnoreCase(columnMetadata.typeName())) {
+            return true;
+        }
+        return matchingDialectColumnType(columnMapping, columnMetadata);
+    }
+
+    /**
+     * Provides dialect-specific logic for matching the column type defined in the {@link SqlColumnMapping}
+     * with the actual column type retrieved from the database metadata ({@link SqlColumnMetadata}).
+     * <p>
+     * This method is called when the default type comparison in {@link #matchingColumnType(SqlColumnMapping, SqlColumnMetadata, Dialect)}
+     * fails to match the column types. Implementations should provide custom logic to handle dialect-specific
+     * type mappings or exceptions.
+     * <p>
+     * The default implementation always returns true, indicating that the column types match. Subclasses
+     * can override this method to provide more specific type matching logic for their supported dialect.
+     *
+     * @param columnMapping  the SQL column mapping from {@link PersistentEntity} field
+     * @param columnMetadata the SQL column metadata from the database
+     * @return true if the column type matches according to the dialect-specific logic, false otherwise
+     */
+    protected boolean matchingDialectColumnType(SqlColumnMapping columnMapping, SqlColumnMetadata columnMetadata) {
+        return false;
+    }
+
+    /**
+     * Checks if the given column metadata represents a VARCHAR column with a size of 36,
+     * typically used to store UUIDs.
+     *
+     * @param columnMetadata the SQL column metadata to check
+     * @return true if the column is a VARCHAR(36), false otherwise
+     */
+    protected static boolean uuidMatchesVarchar(SqlColumnMetadata columnMetadata) {
+        return columnMetadata.type() == Types.VARCHAR && columnMetadata.columnSize() == 36;
+    }
+
+    /**
+     * Checks if the given JDBC type code represents a floating-point data type.
+     *
+     * @param typeCode the JDBC type code to check
+     * @return true if the type code is FLOAT, REAL, or DOUBLE, false otherwise
+     */
+    protected static boolean isFloatOrRealOrDouble(int typeCode) {
+        return switch (typeCode) {
+            case Types.FLOAT, Types.REAL, Types.DOUBLE -> true;
+            default -> false;
+        };
+    }
+
+    /**
+     * Validates a single column mapping definition against its corresponding metadata from the database.
+     *
+     * @param columnMapping     the SQL column mapping from {@link PersistentEntity} field to validate
+     * @param columnMetadata    the SQL column metadata from the database to compare against
+     * @param dialect           the dialect of the schema where column belongs
+     * @param tableName         the name of the table where column is stored
+     * @throws SchemaValidationException when the expected column does not match the actual column metadata
+     */
+    private void validateColumn(SqlColumnMapping columnMapping, SqlColumnMetadata columnMetadata,
+                                       Dialect dialect, String tableName) {
+        if (StringUtils.isNotEmpty(columnMapping.getDefinition())) {
+            // Don't compare columns with custom SQL definition
+            // and let user be responsible for mapping of that field
+            return;
+        }
+        if (matchingColumnType(columnMapping, columnMetadata, dialect)) {
+            return;
+        }
+        throw new SchemaValidationException(String.format("Schema validation failed. Column [%s] in table [%s] of type [%s] is mapped to [%s].",
+            columnMetadata.name(), tableName, columnMetadata.typeName(), columnMapping.getDbType()));
+    }
+
+    private static boolean matchingColumnTypes(SqlDbType dbType, int typeCode) {
+        int mappedTypeCode = dbType.getType();
+        return mappedTypeCode == typeCode
+            || isCompatibleIntegralType(mappedTypeCode, typeCode)
+            || isNumericOrDecimal(mappedTypeCode) && isNumericOrDecimal(typeCode)
+            || isFloatOrRealOrDouble(mappedTypeCode) && isFloatOrRealOrDouble(typeCode)
+            || isVarcharType(mappedTypeCode) && isVarcharType(typeCode)
+            || isVarbinaryType(mappedTypeCode) && isVarbinaryType(typeCode)
+            || isEnumType(mappedTypeCode) && isVarcharType(typeCode);
+        // Add more checks/fallbacks during testing and/or reported issues
+    }
+
+    private static boolean isCompatibleIntegralType(int typeCode1, int typeCode2) {
+        return switch (typeCode1) {
+            case Types.TINYINT ->
+                typeCode2 == Types.TINYINT
+                    || typeCode2 == Types.SMALLINT
+                    || typeCode2 == Types.INTEGER
+                    || typeCode2 == Types.BIGINT;
+            case Types.SMALLINT ->
+                typeCode2 == Types.SMALLINT
+                    || typeCode2 == Types.INTEGER
+                    || typeCode2 == Types.BIGINT;
+            case Types.INTEGER ->
+                typeCode2 == Types.INTEGER
+                    || typeCode2 == Types.BIGINT;
+            default -> false;
+        };
+    }
+
+    private static boolean isNumericOrDecimal(int typeCode) {
+        return switch (typeCode) {
+            case Types.NUMERIC, Types.DECIMAL ->  true;
+            default -> false;
+        };
+    }
+
+    private static boolean isVarcharType(int typeCode) {
+        return switch (typeCode) {
+            case Types.VARCHAR, Types.LONGVARCHAR, Types.NVARCHAR, Types.LONGNVARCHAR -> true;
+            default -> false;
+        };
+    }
+
+    private static boolean isVarbinaryType(int typeCode) {
+        return switch (typeCode) {
+            case Types.VARBINARY, Types.LONGVARBINARY -> true;
+            default -> false;
+        };
+    }
+
+    private static boolean isEnumType(int typeCode) {
+        return typeCode == SqlDbType.ENUM.getType();
+    }
+}
